move bookmark function from utils to service/database
